requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
python-telegram-bot==22.2
httpx==0.28.1
httpcore==1.0.9

